- refactored ancilary components to components folder - added assets folder - Made strides towards Holly's Design language 	- header 	- business-list
UpNext - MOAR css, add ng2boostrap directives, extend functionality of business list, Add Footer, and media assets
