+++ app/components/businessType-list.component.html
@@ -1,12 +1,16 @@
-<section class="container">
+<section class="bp-list-container">
 <div class="row">
-    <div class="col-md-6 selected-business-category-column" *ngIf="selectedbusinessCategory">
+    <div class="col-md-4 selected-business-category-column" *ngIf="selectedbusinessCategory">
+        <div class="bp-list-category-header">REFINE YOUR RESULTS</div>
+        <div  class="bp-list-category-subheader">Filter by Category</div>
         <ul>
             <li class="business-category-row" *ngFor="let businessCategory of finalBusinessCategories" (click)="onSelect(businessCategory)" [class.selected]="businessCategory === selectedbusinessCategory">{{businessCategory.name}}
             </li>
         </ul>
     </div>
-    <div class="col-md-6 selected-business-types-permits-column" *ngIf="selectedbusinessCategory">
+    
+    <div class="col-md-8 selected-business-types-permits-column" *ngIf="selectedbusinessCategory">
+    <div class="bp-list-permits-header">PERMITS</div>
         <div>
             <h2 [@flyInOut]="selectedbusinessCategory">
         {{selectedbusinessCategory.name | uppercase}}
