+++ app/components/header.component.html
@@ -1,0 +1,44 @@
+<div class="page-header bp-header">
+  <nav class="navbar navbar-default navbar-static-top bp-nav">
+    <div class="container-fluid bp-nav-container">
+      <!-- Brand and toggle get grouped for better mobile display -->
+    <div class="navbar-header">
+      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
+        <span class="sr-only">Toggle navigation</span>
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>
+      </button>
+      <a class="navbar-brand bp-brand" href="#"></a>
+    </div>
+
+      <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+      <p class="navbar-text bp-navlink-text" ><a href="#" class="navbar-link bp-navlink">Permit Locator</a></p>
+      <ul class="nav navbar-nav navbar-right">
+       <li> <p class="navbar-text navbar-right bp-navlink-text"><a href="#" class="navbar-link bp-navlink">ABOUT</a></p></li>
+        <p class="navbar-text navbar-right bp-navlink-text"><a href="#" class="navbar-link bp-navlink">CONTACT</a></p>      
+        </ul>
+      </div>     
+
+      <hr class="bp-nav-hr">
+    </div>  
+  </nav>
+    <div class="row">
+      <div class="col-sm-4 col-sm-offset-4">
+        <p class="bp-header-text"> Find permits, licenses, and forms that apply to your business by using the search field and filtering by type</p>
+      </div>
+    </div>
+
+    <div class="row">
+     <div class="col-sm-4 col-sm-offset-4">
+    <div class="input-group">
+      <input type="text" class="form-control bp-search" placeholder="Search for...">
+      <span class="input-group-btn">
+        <button class="btn btn-default bp-search" type="button">Go!</button>
+      </span>
+    </div>
+  </div>
+  </div>
+
+
+</div>
