+++ app/components/header.component.ts
@@ -1,7 +1,7 @@
 ﻿import { Component } from '@angular/core';
 @Component({
     selector: 'header',
-    templateUrl: './header.component.html'
+    templateUrl: './header.component.html',
 })
 export class HeaderComponent {
 }
