--- app/header.component.css
@@ -1,3 +1,0 @@
-/*.bp-nav-header{
-	margin-bottom: 20px;
-}*/
