+++ typings/globals/jquery/index.d.ts
@@ -3221,8 +3221,7 @@ interface JQuery {
     queue(queueName: string, callback: Function): JQuery;
 }
 declare module "jquery" {
-    export var $:JQueryStatic;
+    export = $;
 }
-
-// declare var jQuery: JQueryStatic;
-// declare var $: JQueryStatic;
+declare var jQuery: JQueryStatic;
+declare var $: JQueryStatic;
