+++ typings/index.d.ts
@@ -1,4 +1,3 @@
-/// <reference path="globals/angular-protractor/index.d.ts" />
 /// <reference path="globals/core-js/index.d.ts" />
 /// <reference path="globals/jasmine/index.d.ts" />
 /// <reference path="globals/jquery/index.d.ts" />
